API_Smart_Park/requirements.txt
fastapi~=0.110.1
uvicorn[standard]
boto3
pydantic
pydantic[email]
PyJWT
